requirements.txt
openai-whisper
fastapi 
uvicorn 
python-multipart
imageio[ffmpeg]
librosa 
soundfile
